+++ flowmachine/flowmachine/core/dummy_query.py
@@ -20,13 +20,11 @@ class DummyQuery(Query):
 
     def __init__(self, dummy_param):
         self.dummy_param = dummy_param
+        super().__init__()
 
     @property
     def query_id(self):
-        # Prefix the usual query_id hash with 'dummy_query' to make it obvious
-        # that this is not a regular query.
-        md5_hash = super().query_id
-        return f"dummy_query_{md5_hash}"
+        return f"dummy_query_{self.dummy_param}"
 
     def _make_query(self):
         return "SQL_of_dummy_query"