+++ flowmachine/tests/server/test_query_logging.py
@@ -3,6 +3,8 @@
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 import json
+
+import pytest
 import sys
 from logging import getLogger
 
@@ -11,7 +13,8 @@ from flowmachine.core.server.server import get_reply_for_message
 from flowmachine.core import Query
 
 
-def test_query_run_logged(json_log, server_config):
+@pytest.mark.asyncio
+async def test_query_run_logged(json_log, server_config):
     # Local import so pytest can capture stdout
     logger = getLogger("flowmachine.query_run_log")
     logger.handlers[0].stream = sys.stdout  # Reset log stream for capsys
@@ -26,7 +29,7 @@ def test_query_run_logged(json_log, server_config):
     Query.redis.get.return_value = (
         b"known"  # Mock enough redis to get to the log messages
     )
-    reply = get_reply_for_message(
+    reply = await get_reply_for_message(
         msg_str=json.dumps(msg_contents), config=server_config
     )